refactor(index.ts): 👷 Use direct imports
