+++ app/personal-projects/page.tsx
@@ -3,7 +3,8 @@ import ProjectsLayout from '@/components/layout/ProjectsLayout'
 import PageNavigation from '@/components/page-navigation/PageNavigation'
 import { ProjectSection } from '@/components/projects/overview-page/ProjectSection'
 import BreadCrumbs from '@/components/shared/Breadcrumbs'
-import { projectsPersonalNext, projectsPersonalReact } from '@/data/projects/personal/projects-overview'
+import { projectsPersonalNext } from '@/data/projects/personal/projects-overview/personalNext'
+import { projectsPersonalReact } from '@/data/projects/personal/projects-overview/personalReact'
 import { TEXT } from '@/localization/english'
 import { DATA_TEST_IDS } from '@/utils/constants/ids/dataTestIds'
 import { ID } from '@/utils/constants/ids/elementIds'
