--- components/icons/index.ts
@@ -1,8 +1,0 @@
-import { IconArrow } from './IconArrow'
-import { IconCaretRight } from './IconCaretRight'
-import { IconCheck } from './IconCheck'
-import { IconClose } from './IconClose'
-import { IconHamburger } from './IconHamburger'
-import { IconHome } from './IconHome'
-
-export { IconArrow, IconCaretRight, IconCheck, IconClose, IconHamburger, IconHome }
