+++ components/layout/project-page/HeaderSection.tsx
@@ -1,4 +1,4 @@
-import { IconArrow } from '@/components/icons'
+import { IconArrow } from '@/components/icons/IconArrow'
 import Heading1 from '@/components/shared/Heading1'
 import Heading2 from '@/components/shared/Heading2'
 import { getGoBackLinkID } from '@/utils/helpers/getGoBackLink'
