--- data/projects/personal/projects-overview/index.ts
@@ -1,2 +1,0 @@
-export * from './personalNext'
-export * from './personalReact'
