--- data/projects/work/projects-overview/index.ts
@@ -1,5 +1,0 @@
-export * from './workFrontEnd'
-export * from './workLocalization'
-export * from './workQA'
-export * from './workReact'
-export * from './workWordpress'
